refactor(web): decompose the graph page
Move graph state, projection, controls, canvas rendering, and detail views
into cohesive modules behind a thin route boundary.

Preserve the existing query contract, D3 behavior, responsive layout, and
interactive controls while adding focused tests for the extracted seams.
